major canvas refactor
