+++ src/app/features/3d-volumes/components/settings/components/texture-settings/texture-settings.component.ts
@@ -26,9 +26,6 @@ export class TextureSettingsComponent implements OnInit {
   public inputTextureSizeX: CustomInput;
   public inputTextureSizeY: CustomInput;
 
-  private shaderUvConfig$: Observable<ShaderConfig> = this.canvasService.getShaderConfig()
-    .pipe(takeUntilDestroyed());
-
   private outputResolution: IVector2 = this.canvasService.getOutputResolution();
 
   constructor(
@@ -51,7 +48,6 @@ export class TextureSettingsComponent implements OnInit {
     ];
 
     this.maxTextureSize = getMaxTextureSize();
-    this.handleUvConfigChanges();
   }
 
   public updateOutputResolutionX(value: string | number): void {
@@ -72,17 +68,6 @@ export class TextureSettingsComponent implements OnInit {
     this.canvasService.updateOutputResolution(
       { x: this.outputResolution.x, y: value }
     );
-  }
-
-  private handleUvConfigChanges(): void {
-    this.shaderUvConfig$.subscribe((config: ShaderConfig) => {
-      for (const [name, value] of Object.entries(config)) {
-        const slider = this.getSliderByUniformName(name);
-        if (slider) {
-          slider.value = value;
-        }
-      }
-    });
   }
 
   public onSliderChange(slider: Slider): void {
